change control_id to uint32_t
it is 4 bytes the arturia v3 protocol
preset id (0x08 global)
parameter id
control id 1
control id 2

+++ src/sc-arturia-book.c
@@ -113,7 +113,7 @@ sc_arturia_book_class_init (ScArturiaBookClass *klass)
 }
 
 int
-sc_arturia_book_read_control (ScArturiaBook *self, uint16_t control_id, uint8_t *val)
+sc_arturia_book_read_control (ScArturiaBook *self, uint32_t control_id, uint8_t *val)
 {
   ScArturiaBookClass *klass;
   ScArturiaBookPrivate *priv;
@@ -127,7 +127,7 @@ sc_arturia_book_read_control (ScArturiaBook *self, uint16_t control_id, uint8_t 
 }
 
 int
-sc_arturia_book_write_control (ScArturiaBook *self, uint16_t control_id, uint8_t val)
+sc_arturia_book_write_control (ScArturiaBook *self, uint32_t control_id, uint8_t val)
 {
   ScArturiaBookClass *klass;
   ScArturiaBookPrivate *priv;
