+++ src/sc-preferences-group.h
@@ -13,6 +13,6 @@ struct _ScPreferencesGroupClass
   AdwPreferencesGroupClass parent_class;
 };
 
-uint16_t sc_preferences_group_get_control_id_offset (ScPreferencesGroup *self);
+uint32_t sc_preferences_group_get_control_id_offset (ScPreferencesGroup *self);
 
 G_END_DECLS
